Rename donationAmounts
